feat: add `jobs.getExportUrls` utility
This adds a utility method which allows library consumers to turn a job into a list of file results.

Contrary to what was suggested in #91, this does not add a class `Job` which would have to be instantiated. Therefore, it does not require a hydration of all API call results. Currently, all interactions with jobs and tasks happen through `cloudconvert.jobs` and `cloudconvert.tasks`, so doing this would work, but not be very consistent.

Closes #91.